requirements.txt
pyserial==3.4
click==7.1.2
influxdb==5.3.1
python-dateutil==2.8.1
pytz==2020.4